requirements.txt
mysqlclient==1.3.6
Django==1.8.2
django-timezone-field
django-crispy-forms
django-suit
